+++ doc/release_notes.rst
@@ -18,6 +18,7 @@ Release Notes
 
 * feat: Add options for carrier specific load shedding and load sinks configurable via `load_shedding` and `load_sinks` respectively (https://github.com/PyPSA/pypsa-eur/pull/2105).
 
+* Update heat source handling in `prepare_sector_network` and introduce preheating of heat sources for more realistic system integrations (https://github.com/PyPSA/pypsa-eur/pull/1893).
 PyPSA-Eur v2026.02.0 (18th February 2026)
 =========================================